#5473 - Add HotWaterEquipment convenience methods to SpaceType (don't like them)
Note that I dislike the getters such as hotWaterEquipmentPowerPerFloorArea, hotWaterEquipmentPowerPerPerson because they just query the equipment instnaces on the space type and work only if they are ALL PerFloorArea/PerPerson.

I think it would be more useful if they looped on SpaceType::spaces() and would sum the hotWaterEquipmentPowerPerPerson (times the multiplier).

But I'm matching the other SpaceLoadInstances (GasEquipment, ElecttricEquipment, etc)